Add some tests, remove old template tags
